+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Implemented `version` command to show the version of the tm-catalog-cli
 - Implemented autocompletion for most flags and arguments for the shell autocompletion script
+- Added an optional flag to fetch to restore original external id to the fetched TM
 
 ### Changed